+++ .storybook/__conf__/mochaMockConfig.js
@@ -1,27 +1,26 @@
-import {storiesOf, action, linkTo, specs, describe, it} from "../__mocks__/facade-mocha";
+import {storiesOf, action, linkTo, specs, describe, it,} from "../__mocks__/facade-mocha";
 global.storiesOf = storiesOf;
 global.action = action;
 global.linkTo = linkTo;
 global.specs = specs;
 global.describe = describe;
 global.it = it;
+global.after = after;
+global.before = before;
+global.beforeEach = beforeEach;
+global.afterEach = afterEach;
 
-require('babel-register')();
+import { jsdom } from 'jsdom';
 
-var jsdom = require('jsdom').jsdom;
-
-var exposedProperties = ['window', 'navigator', 'document'];
-
-global.document = jsdom('');
-global.window = document.defaultView;
-Object.keys(document.defaultView).forEach((property) => {
-  if (typeof global[property] === 'undefined') {
-    exposedProperties.push(property);
-    global[property] = document.defaultView[property];
+/**
+ * Mocking browser-like DOM
+ */
+global.document = jsdom('<!doctype html><html><body></body></html>', {
+  headers: {
+    'User-Agent':
+    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_6_7)' +
+    ' AppleWebKit/534.24 (KHTML, like Gecko) Chrome/11.0.696.71 Safari/534.24'
   }
 });
-
-global.navigator = {
-  userAgent: 'node.js'
-};
-
+global.window = document.defaultView;
+global.navigator = global.window.navigator;