+++ .storybook/__mocks__/facade.js
@@ -31,4 +31,11 @@ export const snapshot = (name, story) => {
 
 export const describe = jasmine.currentEnv_.describe;
 export const it = jasmine.currentEnv_.it;
+export const beforeEach = jasmine.currentEnv_.beforeEach;
+export const afterEach = jasmine.currentEnv_.afterEach;
+export const xit = jasmine.currentEnv_.xit;
+export const xdescribe = jasmine.currentEnv_.xdescribe;
+export const fit = jasmine.currentEnv_.fit;
+export const after = () => {};
+export const before = () => {};