+++ .storybook/__tests__/sample.ci.jest.stories.js
@@ -1,10 +1,12 @@
 import React from "react";
-import {storiesOf, action, describe, it, specs} from "../facade";
+import {storiesOf, action, describe, it, specs,
+beforeEach, before, after, afterEach, xdescribe,
+fit, xit} from "../facade";
 import {mount} from "enzyme";
 import expect from "expect";
 
 
-const stories = storiesOf('Button - CI Sample', module);
+const stories = storiesOf('Button - CI JEST Sample', module);
 
 stories.add('Hello World', function () {
   const helloWorldStory =
@@ -13,10 +15,24 @@ stories.add('Hello World', function () {
     </button>;
 
   specs(() => describe('Hello World', function () {
+    let output;
+    beforeEach(function() {
+      console.log('BEFORE EACH');
+      output = mount(helloWorldStory);
+    });
+
+    afterEach(function() {
+      console.log('AFTER EACH');
+    });
+
     it('Should have the Hello World label', function () {
-      let output = mount(helloWorldStory);
       expect(output.text()).toContain('Hello World');
     });
+
+    it('Should have the Hello World label', function () {
+      expect(output.text()).toContain('Hello World');
+    });
+
   }));
   return helloWorldStory;
 });