+++ .storybook/__tests__/sample.ci.mocha.stories.js
@@ -2,8 +2,7 @@ import React from "react";
 import {mount} from "enzyme";
 import expect from "expect";
 
-
-const stories = storiesOf('Button', module);
+const stories = storiesOf('Button - CI MOCHA Sample', module);
 
 stories.add('Hello World', function () {
   const helloWorldStory =
@@ -12,15 +11,32 @@ stories.add('Hello World', function () {
     </button>;
 
   specs(() => describe('Hello World', function () {
-    it('Should have the Hello World label', function () {
-      let output = mount(helloWorldStory);
-      expect(output.text()).toContain('Hello Wrld');
+    let output;
+    beforeEach(function() {
+      console.log('BEFORE EACH');
+      output = mount(helloWorldStory);
+    });
+
+    before(function() {
+      console.log('BEFORE');
+    });
+
+    afterEach(function() {
+      console.log('AFTER EACH');
+    });
+
+    after(function() {
+      console.log('AFTER ');
     });
 
     it('Should have the Hello World label', function () {
-      let output = mount(helloWorldStory);
       expect(output.text()).toContain('Hello World');
     });
+
+    it('Should have the Hello World label', function () {
+      expect(output.text()).toContain('Hello World');
+    });
+
   }));
   return helloWorldStory;
 });