+++ src/index.js
@@ -4,4 +4,6 @@ export const PANEL_ID = `${ADDON_ID}/specifications-panel`;
 export const EVENT_ID = `${ADDON_ID}/specifications-event`;
 
 export { register } from './manager';
-export { specs, describe, it } from './preview';
+export { specs, describe, it,
+  beforeEach, afterEach, after, before,
+xit, fit, xdescribe} from './preview';