+++ CHANGELOG.md
@@ -8,3 +8,14 @@
 
 #2: log specs failure in the console
 #3: improve facade.js to avoid modifying your stories declaration when adding ci
+
+### v1.0.14
+
+* change specs display to improve failures readability
+
+### v1.0.15
+
+* add hooks and jest/mocha specifics functions :
+  * before, after, beforeEach, afterEach
+  * jest : xit, fit, xdescribe
+  * mocha : describe.only, describe.skip, it.only, it.skip